Interact with one item
Set up raycast to only react to objects with object type "interactable". This removes the issue where an object can continue to be attempted interactions even when already busy (player faces crop when not supposed to).

Fixed error in the console where it tried to overlap with objects that got destroyed mid-interaction (changed when collisions are turned back on) and a second hidden error log caused when the player picks up and destroys an item before it has finished teleporting rest of the object to static mesh.