Hoe and Watering Can
Set up temporary inventory equipment system that allows the player to click on the num keys and, if the player currently has that item in their inventory, equips a specific item.

The player can now equip a Hoe, watering can, or seed and then interact with the soil to plant

Tidied up code to remove unnecessary code and breaking player blueprint into functions so that reusable code only requires one node and making code easier to read with inclusion of comments.