AI Quest System & new animations
Finished quest system so now it can the quest type can be specified (fetch quest or not).

Items required for a quest, now get removed from the player's inventory upon completion and a quest can only be completed with the required number of items.

Also added new animations to slime NPC.