Navigation Update
Altered main buildings for grasslands so that navigation mesh no longer generates inside of models. Also made it so that the trigger boxes that enable the AI to 'fake enter' buildings, no longer affect the navmesh. Created a blueprint version of the fence so that collision could be added to it and update the navmesh correctly.