Quest system
Fixed bugs with quest system. The player now requires wood once again to be able to build the bridge. Other npcs will not attempt to give the player a quest if the active quest has not been completed. Changed quest text for the build bridge quest to give more context. Changed the default NPC widget text to be 'Good day!'. The active and completed quests are now stored on a component on the player.